refactor: Move providers outside of the applications-api package
